dataset/translit/requirements.txt
ai4bharat-transliteration
fastapi
uvicorn
pydantic
dotenv
